refactor(harness): promote FsDirEntry/FsListResponse into shared/types.ts
server/fs.ts and web/lib/api.ts each declared identical local copies of
these types since the fs-browsing endpoint predated the shared contract.
Point both at src/shared/types.ts instead so the contract stays the
single source of truth.
